投手総合成績のイニングを小数点表示に変更

+++ eikan/models/players_pitcher.py
@@ -22,9 +22,9 @@ class PlayersPitcher(models.Model):
         editable=False,
     )
 
-    innings_pitched = models.PositiveSmallIntegerField(
+    innings_pitched = models.FloatField(
         verbose_name="イニング",
-        default=0,
+        default=0.0,
         editable=False,
     )
